feat_api(be): Implemented Service to calculate cost of registers based on the dynamic formula defined for company

+++ backend/ElectroManage.Application/Features/Register/Command/Post/CreateRegisterCommand.cs
@@ -3,6 +3,6 @@
 public record CreateRegisterCommand : ICommand<CreateRegisterResponse>
 {
     public long CompanyId { get; set; }
-    public decimal Consumption { get; set; }
+    public double Consumption { get; set; }
     public DateTime Date { get; set; }
 }
